refactor CSV downloads to support async data
