+++ .vscode/project-words.txt
@@ -24,6 +24,7 @@ darkgray
 darknavy
 darkpink
 darkpurple
+datas
 datetime
 DCMO
 dedup
