+++ packages/ui/package.json
@@ -44,7 +44,7 @@
     "chromatic": "^10.2.0",
     "framer-motion": "^10.17.6",
     "lodash": "4.17.21",
-    "react-csv": "2.2.2",
+    "react-csv-downloader": "^3.1.0",
     "react-dom": "^18.2.0",
     "react-focus-lock": "^2.9.6",
     "react-remove-scroll": "^2.5.7",
