+++ packages/ui/src/components/Link/DownloadCsv.tsx
@@ -1,13 +1,13 @@
 import React from "react";
-import { CSVLink } from "react-csv";
+import CsvDownloader, { ICsvProps } from "react-csv-downloader";
 
 import ButtonLinkContent from "../ButtonLinkContent/ButtonLinkContent";
 import getButtonStyles from "../../utils/button/getButtonStyles";
 import type { LinkProps } from "./Link";
 
 interface CsvHeader {
-  key: string;
-  label: string;
+  id: string;
+  displayName: string;
 }
 
 export interface DownloadCsvProps
@@ -16,7 +16,7 @@ export interface DownloadCsvProps
     "color" | "mode" | "block" | "type" | "icon" | "fontSize"
   > {
   headers: CsvHeader[];
-  data: Record<string, string>[];
+  data: ICsvProps["datas"];
   fileName: string;
   children: React.ReactNode;
 }
@@ -34,18 +34,18 @@ const DownloadCsv = ({
   ...rest
 }: DownloadCsvProps) => {
   return (
-    <CSVLink
+    <CsvDownloader
+      data-h2-cursor="base(pointer)"
       {...getButtonStyles({ mode, color, block })}
       {...rest}
-      target="_blank"
-      headers={headers}
-      data={data}
+      columns={headers}
+      datas={data}
       filename={fileName}
     >
       <ButtonLinkContent mode={mode} icon={icon} fontSize={fontSize}>
         {children}
       </ButtonLinkContent>
-    </CSVLink>
+    </CsvDownloader>
   );
 };
